misc/gemini_competition/requirements.txt
firebase-admin
python-dotenv
google-generativeai
